+++ lib/components/loading/loading_overlay.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:pyc/common/constants/constants.dart';
-import 'package:shimmer/shimmer.dart';
 
 class LoadingOverlay extends StatelessWidget {
   final Widget child;
