+++ lib/screens/index/index_screen.dart
@@ -2,13 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:pyc/common/constants/constants.dart';
 import 'package:pyc/components/loading/loading_overlay.dart';
+import 'package:pyc/controllers/calendar/index_calendar_controller.dart';
 import 'package:pyc/controllers/notice/notice_controller.dart';
 import 'package:pyc/controllers/user/fetch_me_controller.dart';
 import 'package:pyc/screens/index/components/appbar/index_appbar.dart';
+import 'package:pyc/screens/index/components/content/index_calendar_list.dart';
 import 'package:pyc/screens/index/components/content/index_notice_list.dart';
 import 'package:pyc/screens/index/components/drawer/index_drawer.dart';
 import 'package:pyc/screens/index/components/index_attendance.dart';
-import 'package:pyc/screens/index/components/index_content_card.dart';
 import 'package:pyc/screens/index/components/index_user_profile.dart';
 import 'package:pyc/screens/index/components/index_user_search.dart';
 import 'package:pyc/screens/index/components/layout/labeled_content.dart';
@@ -97,20 +98,15 @@ class IndexScreen extends StatelessWidget {
             LabeledContent(
               title: '일정 안내',
               goContent: () {},
-              child: Column(
-                children: [
-                  IndexContentCard(
-                    avatarChild: const Icon(
-                      Icons.edit_calendar_outlined,
-                      size: kDefaultValue * 1.75,
-                      color: kTextWhiteColor,
-                    ),
-                    title: '2022년 12월 연말 전도축제',
-                    content: '시작일 | 22.12.25',
-                    goTo: () {},
+              child: GetBuilder<IndexCalendarController>(
+                builder: (controller) => LoadingOverlay(
+                  isLoading: controller.isLoading,
+                  child: IndexCalendarList(
+                    cardCount: cardCount,
+                    rows: controller.rows,
+                    count: controller.count,
                   ),
-                  kHalfHeightSizeBox,
-                ],
+                ),
               ),
             ),
             // kHeightSizeBox,
